Initial commit of the Prometheus/OpenTelemetry compatibility spec
This is an early draft that outlines the goals and expectations
from OpenTelemetry in order to provide Prometheus support. The
work is in the early stages, behavioral expectations from the
collector and the libraries might change or be expanded in the
future. In the long term, we will graduate this doc to the
OpenTelemetry compatibility specs under the opentelemetry-specification
repo.